pyproject.toml
[project]
name = "dohome_rgb"
authors = [
    {name = "Mikhael Khrustik", email = "misha@myrt.co"},
]
description = "DoHome lights Home Assistant integration"
readme = "README.md"
requires-python = ">=3.12"
keywords = ["home-assistant", "dohome"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "homeassistant",
    "pylint==4.0.5",
    "ruff==0.15.6",
    "dohome-api==2.2.1",
]
version = "v0.0.0"

[tool.ruff]
lint.ignore = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
lint.select = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.basedpyright]
venvPath = "."
venv = ".venv"
reportSelfClsParameterName = false
reportUnannotatedClassAttribute = false
reportConstantRedefinition = false
reportExplicitAny = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportAny = false
